Move from only identifying accounts by their names to temporary IDs.
A fundamental issue with pizauth is that configuration reloading brings
the opportunity for subtle security bugs. For example, pizauth might be part
of the way through obtaining a token for an account "x" when the user
reloads the config and changes part of "x"s settings (e.g. `token_uri`).
This could cause pizauth to leak a token generated from one system to
another entirely different system.

This commit is the first in a multi-stage approach to stop this
happening. This commit introduces `CTGuardAccountId` which are opaque
account identifiers. They are only fully valid while ct_lk is held: when
ct_lk is dropped, a `CTGuardAccountId` cannot be used to lookup a
`TokenState`, but it can be used to compare it to a new
`CTGuardAccountId`. This will allow us to determine "is a new
`CTGuardAccountId` the same as an old `CTGuardAccountId`" where "the
same" will mean roughly "the underlying `Account`s compare equal". If we
have to drop the lock and reacquire it, this allows us to determine
whether we're really operating on the account we were before dropping
the lock.

Because this security property is something we care about so much, we
had extra tracking information to make it impossible to reuse a
`CTGuardAccountId` after its parent `CTGuard` has dropped: trying to
reuse it will cause a `panic`. We use `Arc`'s reference counting with
weak references as a simple way of tracking this.

This commit is more-or-less the minimal change necessary to change the
internal API. There are still quite a few parts that use account name
strings when they should be using `CTGuardAccountId`s. We can change
those gradually.