Reset the tokenstate to Empty for accounts that have changed on reload.
This is an important security property: if the user changes an account's
details in `pizauth.conf` we can't know if the altered account is really
"the same" or not. So if any aspects of the account alter, we ensure the
tokenstate is set to `Empty`. In other words, reloading config can cause
you to have to reauthenticate an account. This is a more acceptable
outcome than leaking tokens!