Add a tokenstate version to `CTGuardAccountId`.
Previously a `CTGuardAccountId` only failed to revalidate if the
underlying account changed. This commit makes it change if the
`TokenState` has changed in anyway, replacing several ad-hoc (and not
entirely correct) checks along these lines. This means that if two
threads race on changing a `TokenState`, one of them definitively wins
and makes progress, rather than the previous situation where they could
become entangled in odd ways.

In order for this to work, we add a new `TokenStateVersion` struct.
Every time the tokenstate changes, the version is incremented. A
`CTGuardAccountId` keeps a copy of this version number which it can then
use for later equality checks. When a tokenstate is updated, we also
need to receive a new `CTGuardAccountId` in order that its tokenstate
version is upto date.