Don't endlessly spin retrying after a refresh failure.
Refreshing can easily fail for temporary reasons (e.g. lack of an
internet connection). Previously this caused pizauth to spin endlessly.
This commit introduces a `retry_refresh_interval` which sets a minimum
interval to try refreshing. This is set to 40s so that the (new) default
`refresh_before_expiry` of 90s means we can realistically try 3 times
before a token actually expires.