Rename `auth_cmd` to `auth_notify_cmd` and make it optional.
Now that `pizauth show` (and `refresh`) show authorisation URLs by
default, many users might not feel the need to use e.g. `notify-send` or
the like. That means that `auth_cmd` should be optional.

This commit also renames the setting because we are soon going to add
other notification command kinds.