Remove `-c` from all but the server.
This simplifies the interface as none of the other sub-commands (except,
confusingly, `refresh` and `reload`) made meaningful use of this flag.

`reload` allowed you to switch configuration files which seemed like a
good idea at the time, but I now think is unnecessarily confusing: if
you want to switch config files, shut pizauth down, and start the server
with a new configuration file.

`refresh` read the configuration file from disk and used that to refresh
accounts if none were specified. This could lead to confusing behaviour
if the config file on disk specified different accounts than in the
running server. This commit also simplifies the `refresh` command so
that it mandates a single account name to be specified.