Add experimental support for `unveil` on OpenBSD.
`unveil()` is a tricky API in the sense that one can never be quite sure
if a library will try and use a non-`unveil`ed file. This commit "seems
to work", but it will take me a while longer before I'm confident that
all the necessary files are `unveil`ed.