Introduce a fallback expiry time.
Although I haven't encountered such a case, it's likely that one day
that pizauth encounters a server which hands out expiry times which are
too long for pizauth to represent. The most likely reason for that is
that the server hands out ludicrously long expiry times (e.g. u64::MAX).
To handle that, this commit introduces a fallback expiry time (currently
44 minutes) used if pizauth can't represent the expiry time.

There is also a rather likely less case where pizauth is being run so
many billenia (is that a word? it is now!) in the future that it runs
out of space in `u64`s. This is unlikely enough that I'm happy to pass
the problem onto future generations.