Be more discriminating about transitory / permanent ureq errors.
There are many ways in which network connections can go wrong, some
transient, some permanent. This commit more carefully classifies some
return results from ureq as permanent errors. Because it's an external
library, it's difficult to guarantee that exactly the right set are
chosen, but this is at least better than before.