Add refresh_warn_{cmd|interval}.
pizauth differentiates between transient and permanent errors when
refreshing but what happens if transient errors occur too long? Before
this commit, pizauth could try endlessly to unsuccessfully refresh a token.

This commit introduces refresh_warn_interval which will raise a warning
if refreshing does not succeed for a given period of time (defaulting to
60 minutes). Users can then run an arbitrary command when refreshing has
failed for that length of time. If no such command is specified, pizauth
resets the access token so that the user has to reauthenticate from
scratch: errors tend to be much more obvious to users at this point.

There is an obvious situation where this will be annoying: when an
access token expires while a laptop is suspended. To (mostly) avoid
this, we try refreshing before warning. Will that cause pizauth to
authenticate before networks are back up? Time will tell.