Make sure refreshing is tried at last twice before giving up.
This can easily be triggered after a machine suspend: the access token
expired long ago, and the first refresh attempt is likely to fail as the
network will not be fully back up.

This commit ensures that refreshing has to fail at least twice before
considering that transient errors are permanent. This is, inevitably, an
imperfect proxy for what we really want -- but there is no perfect
proxy, and this has the virtue of being simple and better than what we
had before!