Replace `CTGuardAccountId` and `TokenStateVersion` with `AccountId`.
Previously, `CTGuardAccountId` was, sort of, an "account version". For a
given account version, there were then multiple `TokenStateVersion`s.
The connection between these two was `Arc<Account>` and a `Weak`
reference between `TokenStateVersion` and an `Account`. This seemed like
a good idea at the time but is quite hard to understand and, it turns
out, flawed: `Weak` references cannot be sent between threads (even with
a `Mutex`).

This commit unifies these two version concepts into an `AccountId`.
Everything something about an account changes (its configuration or its
tokenstate) it receives a new, unique, `AccountId`. This is a `u128` to
make it rather unlikely that we exhaust the integer type in any
reasonable run of pizauth. Pleasingly, `AccountId` reduces the code base
in size.

This commit is -- though it may not look like it! -- the minimal change
to keep things both compiling and running. The internal API is now an
unhappy mix of old and new, and will need further tidying up. But by
keeping this commit as small as practical, it's somewhat easier to
convince myself that it's not introducing (too many) new bugs.