Store account IDs and tokenstates in a simpler data structure.
This makes all searches linear, but since it's unlikely that anyone has
more than a handful of accounts, I'm willing to live with that. My guess
is that even with 100s of accounts the speed will be irrelevant. If
performance becomes a problem in the future we can then move to a more
complex data structure.