Change `expect_transient_errors_if` to `not_transient_error_if`.
In retrospect, I was trying to be clever here. Yes,
`expect_transient_errors_if` is easier to write if it expects a non-zero
exit code, but since this is such an unusual way of writing "if", it's
bound to confuse people.

This commit renames the setting to `not_transient_error_if` which is a
bit of a mouthful but seems likely to be more often used correctly.