Only run one refresher thread at a time.
Previously a long-running refresher thread could clash with a second
refresher thread for the same account, preventing the first from ever
noticing if the first had failed "often enough" for a transient error to
be considered permanent.

This commit adds a sort-of lock called `ongoing_refresh` which allows
the refresher thread to avoid trying to run two refresh threads for the
same account. Because we now have timeouts on commands, this should not
stop refreshing making progress, but will allow transient errors to be
better handled when `not_transient_error_if` is set.