Wait for refreshing to have several transitive errors before worrying.
There is a potential "race" condition of sorts when a transitive error
is received and we then run `not_transient_error_if`. If we (or, more
accurately, the user) is unlucky, a very short transitive error could cause
us to run `not_transient_error_if`, discover there's no transitive
error, then invalidate the token.

This commit reuses a previous idea and, and reruns refreshing multiple
times after transitive errors, only running `not_transient_error_if`
after several have failed. The "race" never full goes away, but at some
point the weight of evidence that the transient error is not transient
is high enough that we can consider it to be a permanent error.