Rationalise when we inform the user of errors and so on.
The challenge here is that "error"s occur at two levels: for example,
refreshing a token can lead to an error which means "your refresh token
is no longer invalid" or an error which means "your network connection
is down, so you don't stand a chance of refreshing". Previously, pizauth
somewhat jumbled these together: this commit starts to more cleanly
delineate what, and when, to inform the user.