Don't compare all aspects of an `Account` when reloading the config.
Previously even changing how often you want an account refreshed caused
an `Account` to compare unequal, leading to its corresponding
`TokenState` being set to `Empty`. Not only is this unduly pessimistic,
it doesn't even really make sense, given that changes to the global
setting of the same names didn't cause accounts to be considered
unequal.

This commit removes `PartialEq` and introduces a custom function
`secure_eq` which compares only the parts of an `Account` which are
relevant to security. This function must be carefully considered every
time a new field is added to `Account` or if the semantics of any
existing fields are changed.