Make the HTTP and HTTPS servers optional.
The main use case for this is someone who wants to ensure that only the
HTTPS server is active (i.e. that they can't accidentally connect via
HTTP). But making both of them optional keeps the system more
consistent, and someone might have a use for running the HTTP but not
the HTTPS server!