Make the notifier/refresher wake-up continuously to deal with clock changes.
When waiting to do something (e.g. in the notifier or refresher), we
have the problem that when we ask to be woken up in "X seconds from
now", operating systems do not interpret that as "wake you up in X
seconds of wall-clock time". For example, if a machine is suspended then
resumed, that time may not be counted as "wait time". The impact of
ntp/adjtime and friends is also unclear. There is no portable way for us
to know if any of these things has even happened, so we are left in the
unhappy situation that if a thread knows it has work to do in the
future, it needs to wake itself up every so often to check if -- without
us knowing it! -- the clock has changed underneath it.

There is no universally good value here. We're dealing with a relatively
rare case, so we don't want to keep waking up too regularly, but if we
wait too long, users will become annoyed. 37 seconds is a somewhat
arbitrary choice (though it is, deliberately, a prime number), but any
other choice would be at least as arbitrary.