If we detect a stale access token, initiate refreshing.
For all sorts of clock-shifting reasons (e.g. suspend/resume), we might
realise that an access token has expired. Rather than just telling the
user "it's out of date" and expecting them to realise that they need to
initiate a manual refresh (or wait until another part of pizauth decides
to do so), immediately kick a refresh off. This makes it more likely
that if the user retries very quickly there will then be a fresh access
token.