Add a simple maximum HTTP request size.
A classic way for a client to go wrong is for it to send endless data to
a server. This commit is simple way of ensuring that such a scenario is
less likely to cause pizauth to run out of memory (note though: we don't
control `read_line`, and that doesn't provide any guarantees about
if/when it returns in potentially-OOM scenarios). Typically HTTP
requests that we see are around 1.5KiB, so a 16KiB upper limit seems
very generous as a way of catching malfunctioning clients.